COMMERCE-1560 - talend-definition - Convert ResourcePropertiesHelper to LiferayOutputResourceProperties, convert LiferayResourceProperties to BaseLiferayResourceProperties and LiferayInputResourceProperties, calculate schema automatically after each operation change
