+++ modules/etl/talend/talend-definition/src/main/java/com/liferay/talend/tliferayoutput/TLiferayOutputProperties.java
@@ -14,27 +14,16 @@
 
 package com.liferay.talend.tliferayoutput;
 
-import com.liferay.talend.LiferayBaseComponentDefinition;
 import com.liferay.talend.common.log.DebugUtils;
 import com.liferay.talend.common.schema.SchemaUtils;
 import com.liferay.talend.connection.LiferayConnectionResourceBaseProperties;
-import com.liferay.talend.properties.ExceptionUtils;
-import com.liferay.talend.resource.LiferayResourceProperties;
-import com.liferay.talend.runtime.LiferaySourceOrSinkRuntime;
-import com.liferay.talend.runtime.ValidatedSoSSandboxRuntime;
-
-import java.io.IOException;
+import com.liferay.talend.resource.LiferayOutputResourceProperties;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import javax.ws.rs.HttpMethod;
 
 import org.apache.avro.Schema;
 
@@ -45,29 +34,23 @@ import org.talend.components.api.component.Connector;
 import org.talend.components.api.component.ISchemaListener;
 import org.talend.components.api.component.PropertyPathConnector;
 import org.talend.components.common.SchemaProperties;
-import org.talend.daikon.NamedThing;
-import org.talend.daikon.SimpleNamedThing;
 import org.talend.daikon.avro.AvroUtils;
 import org.talend.daikon.avro.SchemaConstants;
-import org.talend.daikon.exception.TalendRuntimeException;
 import org.talend.daikon.i18n.GlobalI18N;
 import org.talend.daikon.i18n.I18nMessageProvider;
 import org.talend.daikon.i18n.I18nMessages;
-import org.talend.daikon.properties.PresentationItem;
 import org.talend.daikon.properties.ValidationResult;
 import org.talend.daikon.properties.ValidationResultMutable;
 import org.talend.daikon.properties.presentation.Form;
-import org.talend.daikon.properties.presentation.Widget;
 import org.talend.daikon.properties.property.Property;
 import org.talend.daikon.properties.property.PropertyFactory;
 
 /**
  * @author Zoltán Takács
+ * @author Ivica Cardic
  */
 public class TLiferayOutputProperties
 	extends LiferayConnectionResourceBaseProperties {
-
-	public static final String ADD_QUOTES = "ADD_QUOTES";
 
 	public static final List<String> rejectSchemaFieldNames = Arrays.asList(
 		com.liferay.talend.common.schema.constants.SchemaConstants.
@@ -75,65 +58,6 @@ public class TLiferayOutputProperties
 
 	public TLiferayOutputProperties(String name) {
 		super(name);
-	}
-
-	public ValidationResult afterCalculateSchema() throws Exception {
-		if (_log.isDebugEnabled()) {
-			_log.debug("Endpoint: " + resource.endpoint.getValue());
-		}
-
-		ValidatedSoSSandboxRuntime validatedSoSSandboxRuntime =
-			LiferayBaseComponentDefinition.initializeSandboxedRuntime(
-				getEffectiveLiferayConnectionProperties());
-
-		ValidationResultMutable validationResultMutable =
-			validatedSoSSandboxRuntime.getValidationResultMutable();
-
-		if (validationResultMutable.getStatus() ==
-				ValidationResult.Result.ERROR) {
-
-			return validationResultMutable;
-		}
-
-		LiferaySourceOrSinkRuntime liferaySourceOrSinkRuntime =
-			validatedSoSSandboxRuntime.getLiferaySourceOrSinkRuntime();
-
-		try {
-			Action action = operations.getValue();
-
-			Schema endpointSchema =
-				liferaySourceOrSinkRuntime.getEndpointSchema(
-					resource.endpoint.getValue(), action.getMethodName());
-
-			resource.main.schema.setValue(endpointSchema);
-			temporaryMainSchema = endpointSchema;
-
-			_updateOutputSchemas();
-
-			validationResultMutable.setMessage(
-				i18nMessages.getMessage("success.validation.schema"));
-		}
-		catch (IOException | TalendRuntimeException e) {
-			validationResultMutable.setMessage(
-				i18nMessages.getMessage("error.validation.schema"));
-			validationResultMutable.setStatus(ValidationResult.Result.ERROR);
-
-			_log.error("Unable to generate schema", e);
-		}
-
-		refreshLayout(getForm(Form.MAIN));
-
-		return validationResultMutable;
-	}
-
-	public void afterOperations() {
-		if (_log.isDebugEnabled()) {
-			Action action = operations.getValue();
-
-			_log.debug("Selected method: " + action.getMethodName());
-		}
-
-		refreshLayout(getForm(Form.MAIN));
 	}
 
 	@Override
@@ -145,24 +69,6 @@ public class TLiferayOutputProperties
 	public void setupLayout() {
 		super.setupLayout();
 
-		Form mainForm = getForm(Form.MAIN);
-
-		operations.setRequired();
-
-		Widget operationsWidget = Widget.widget(operations);
-
-		operationsWidget.setLongRunning(true);
-		operationsWidget.setWidgetType(Widget.ENUMERATION_WIDGET_TYPE);
-
-		mainForm.addRow(operationsWidget);
-
-		Widget calculateSchemaWidget = Widget.widget(calculateSchema);
-
-		calculateSchemaWidget.setLongRunning(true);
-		calculateSchemaWidget.setWidgetType(Widget.BUTTON_WIDGET_TYPE);
-
-		mainForm.addRow(calculateSchemaWidget);
-
 		Form advancedForm = getForm(Form.ADVANCED);
 
 		advancedForm.addRow(dieOnError);
@@ -173,12 +79,9 @@ public class TLiferayOutputProperties
 		super.setupProperties();
 
 		dieOnError.setValue(true);
-		operations.setValue(null);
 
-		operations.setPossibleValues((List<?>)null);
-		operations.setTaggedValue(ADD_QUOTES, true);
-
-		resource = new ResourcePropertiesHelper("resource");
+		resource = new LiferayOutputResourceProperties(
+			"resource", schemaFlow, schemaReject);
 
 		resource.connection = connection;
 
@@ -203,7 +106,8 @@ public class TLiferayOutputProperties
 					}
 
 					if (schema.equals(SchemaProperties.EMPTY_SCHEMA)) {
-						resource.main.schema.setValue(temporaryMainSchema);
+						resource.main.schema.setValue(
+							SchemaProperties.EMPTY_SCHEMA);
 					}
 				}
 
@@ -220,161 +124,10 @@ public class TLiferayOutputProperties
 		validationResultMutable.setStatus(validationResult.getStatus());
 	}
 
-	public transient PresentationItem calculateSchema = new PresentationItem(
-		"calculateSchema");
 	public Property<Boolean> dieOnError = PropertyFactory.newBoolean(
 		"dieOnError");
-	public Property<Action> operations = PropertyFactory.newEnum(
-		"operations", Action.class);
 	public SchemaProperties schemaFlow = new SchemaProperties("schemaFlow");
 	public SchemaProperties schemaReject = new SchemaProperties("schemaReject");
-
-	/**
-	 * Have to use an explicit class to get the override of afterResource(), an
-	 * anonymous class cannot be public and thus cannot be called via Talend's
-	 * reflection mechanism.
-	 *
-	 * @review
-	 */
-	public class ResourcePropertiesHelper extends LiferayResourceProperties {
-
-		public ResourcePropertiesHelper(String name) {
-			super(name);
-		}
-
-		@Override
-		public ValidationResult afterEndpoint() throws Exception {
-			if (_log.isDebugEnabled()) {
-				_log.debug("Endpoint: " + endpoint.getValue());
-			}
-
-			ValidatedSoSSandboxRuntime validatedSoSSandboxRuntime =
-				LiferayBaseComponentDefinition.initializeSandboxedRuntime(
-					getEffectiveLiferayConnectionProperties());
-
-			ValidationResultMutable validationResultMutable =
-				validatedSoSSandboxRuntime.getValidationResultMutable();
-
-			if (validationResultMutable.getStatus() ==
-					ValidationResult.Result.ERROR) {
-
-				return validationResultMutable;
-			}
-
-			LiferaySourceOrSinkRuntime liferaySourceOrSinkRuntime =
-				validatedSoSSandboxRuntime.getLiferaySourceOrSinkRuntime();
-
-			Set<String> supportedOperations = Collections.emptySet();
-
-			try {
-				supportedOperations =
-					liferaySourceOrSinkRuntime.getSupportedOperations(
-						endpoint.getValue());
-			}
-			catch (TalendRuntimeException tre) {
-				endpoint.setValue(null);
-				operations.setValue(null);
-
-				operations.setPossibleValues((List<?>)null);
-			}
-
-			if (supportedOperations.isEmpty()) {
-				operations.setPossibleValues(Action.Unavailable);
-			}
-			else {
-				Stream<String> operationsStream = supportedOperations.stream();
-
-				List<Action> actions = operationsStream.filter(
-					operation -> {
-						Set<String> availableMethodNames =
-							Action.getAvailableMethodNames();
-
-						return availableMethodNames.contains(operation);
-					}
-				).map(
-					this::_toAction
-				).collect(
-					Collectors.toList()
-				);
-
-				operations.setPossibleValues(actions);
-			}
-
-			operations.setValue(null);
-
-			populateParametersTable(liferaySourceOrSinkRuntime);
-
-			refreshLayout(getForm(Form.MAIN));
-
-			return validationResultMutable;
-		}
-
-		@Override
-		public ValidationResult beforeEndpoint() throws Exception {
-			ValidatedSoSSandboxRuntime validatedSoSSandboxRuntime =
-				LiferayBaseComponentDefinition.initializeSandboxedRuntime(
-					getEffectiveLiferayConnectionProperties());
-
-			ValidationResultMutable validationResultMutable =
-				validatedSoSSandboxRuntime.getValidationResultMutable();
-
-			if (validationResultMutable.getStatus() ==
-					ValidationResult.Result.ERROR) {
-
-				return validationResultMutable;
-			}
-
-			LiferaySourceOrSinkRuntime liferaySourceOrSinkRuntime =
-				validatedSoSSandboxRuntime.getLiferaySourceOrSinkRuntime();
-
-			try {
-				Set<String> endpoints =
-					liferaySourceOrSinkRuntime.getEndpointList(HttpMethod.POST);
-
-				endpoints.addAll(
-					liferaySourceOrSinkRuntime.getEndpointList(
-						HttpMethod.PATCH));
-				endpoints.addAll(
-					liferaySourceOrSinkRuntime.getEndpointList(
-						HttpMethod.DELETE));
-
-				List<NamedThing> endpointsNamedThing = new ArrayList<>();
-
-				endpoints.forEach(
-					endpoint -> endpointsNamedThing.add(
-						new SimpleNamedThing(endpoint, endpoint)));
-
-				if (endpoints.isEmpty()) {
-					validationResultMutable.setMessage(
-						i18nMessages.getMessage("error.validation.resources"));
-					validationResultMutable.setStatus(
-						ValidationResult.Result.ERROR);
-
-					return validationResultMutable;
-				}
-
-				endpoint.setPossibleNamedThingValues(endpointsNamedThing);
-			}
-			catch (Exception e) {
-				return ExceptionUtils.exceptionToValidationResult(e);
-			}
-
-			return null;
-		}
-
-		private Action _toAction(String method) {
-			Stream<Action> actionsStream = Action.getActionsStream();
-
-			return actionsStream.filter(
-				action -> method.equals(action.getMethodName())
-			).findFirst(
-			).orElseThrow(
-				() -> new UnsupportedOperationException(
-					String.format("Unsupported operation: %s.", method))
-			);
-		}
-
-	}
 
 	@Override
 	protected Set<PropertyPathConnector> getAllSchemaPropertiesConnectors(
@@ -437,7 +190,7 @@ public class TLiferayOutputProperties
 
 		schemaFlow.schema.setValue(inputSchema);
 
-		Schema rejectSchema = createRejectSchema(inputSchema);
+		Schema rejectSchema = SchemaUtils.createRejectSchema(inputSchema);
 
 		schemaReject.schema.setValue(rejectSchema);
 	}
