--- modules/etl/talend/talend-definition/src/main/resources/com/liferay/talend/resource/LiferayResourceProperties.properties
@@ -1,11 +1,0 @@
-error.validation.resources=Unable to find any exposed resources. Please check the endpoint URL specified for the server.
-error.validation.schema=Unable to determine schema for the selected endpoint.
-
-form.Advanced.displayName=Advanced
-form.Advanced.title=Advanced
-form.Main.displayName=Main
-form.Main.title=Main
-form.Reference.displayName=Reference
-form.Reference.title=Reference
-
-property.endpoint.displayName=Endpoint
